Prune empty auth fragments from prior bridge when witnessing a new note.
This works, but may need to be altered depending upon how we approach
re-scans. If we want to be able to rewind the tree to any witness
boundary via simple truncation (not rewinding to a checkpoint), this
approach could be problematic.

It might be better to change how checkpoints work. At one point in
development they shared the witness mechanism, but duplicate checkpoints
caused a problem.